Do not stop monitor service on power disconnect
The intents for power connected and disconnected are only sent to
explicitly registered receivers, therefore the service needs to be
running all the time.
This simplifies the state tracking significantly, as there is no need
for a scheduled job depending on the system, which was broken in various
ways. Instead, the monitor service starts at boot and is not stopped
when the power is disconnected. Instead, the battery monitor also tracks
the power connection and notifies the monitor service, which manages the
alarm notification.
The notification is probably broken for now, but will be fixed
separately.