Rework notification management
Rename `AlarmManager' to `AlarmController' to go with the "controller
theme" and resolve the naming clash with the Android framework.
Split the notification controller class into a class for the monitor
service and for the alarm controller, with an abstract base class for
the common code.
Declare the members of the application class as `lateinit var' to
prevent them from accessing a `null Context' upon initialization.